refactor: repository의 인터페이스와 구현체를 분리하는 작업 수행
레포지토리의 구현체로만 구성된 부분을 인터페이스와 구현체로 분리하는 작업을 수행
