+++ src/main/java/gift/Application.java
@@ -1,7 +1,9 @@
 package gift;
 
+import gift.config.ProductConfig;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
 
 @SpringBootApplication
 public class Application {
